move metadata import
